+++ cli/tests/pcluster/utils.py
@@ -11,6 +11,9 @@
 import os
 from copy import deepcopy
 
+from assertpy import assert_that
+
+from pcluster.constants import LAMBDA_VPC_ACCESS_MANAGED_POLICY
 from pcluster.schemas.cluster_schema import ClusterSchema
 from pcluster.utils import load_yaml_dict
 
@@ -46,3 +49,33 @@ def get_resources(
             )
         )
     )
+
+
+def assert_lambdas_have_expected_vpc_config_and_managed_policy(generated_template, expected_vpc_config):
+    resources = generated_template.get("Resources")
+
+    for lambda_function in _get_lambda_functions(resources):
+        role = resources.get(_get_role_name(lambda_function))
+
+        if expected_vpc_config:
+            assert_that(_get_vpc_config(lambda_function)).is_equal_to(expected_vpc_config)
+            assert_that(_get_managed_policy_arns(role)).contains(LAMBDA_VPC_ACCESS_MANAGED_POLICY)
+        else:
+            assert_that(_get_vpc_config(lambda_function)).is_none()
+            assert_that(_get_managed_policy_arns(role)).does_not_contain(LAMBDA_VPC_ACCESS_MANAGED_POLICY)
+
+
+def _get_vpc_config(lambda_function):
+    return lambda_function.get("Properties").get("VpcConfig")
+
+
+def _get_role_name(lambda_function):
+    return lambda_function.get("Properties").get("Role").get("Fn::GetAtt")[0]
+
+
+def _get_lambda_functions(resources):
+    return [res for res in resources.values() if res.get("Type") == "AWS::Lambda::Function"]
+
+
+def _get_managed_policy_arns(role):
+    return {arn.get("Fn::Sub") for arn in role.get("Properties").get("ManagedPolicyArns", [])}